thunderbird helper: split helper function into extra file
To improve the readability of the starting wrapper script let's split
off all the helper stuff like local functions, static strings and some
checks. The plain wrapper script for /u/b/thunderbird get consolidated
and compacted by this.

Also adding the Mate DE to the detected environments we can run on, as
suggested by #855286.
Adding the 'mate' string to the zenity call if detected. We still need
to add some sanity checks if zenity or kdialog available.

Closes: #855286