tb-wrapper*: complete rework and moving over for symlinking
The two helper scripts got completely restructured to fit into the idea
we do a symlinking of .thunderbird to .icedove in the users home.
To goal this the various steps got break of into subfunctions that are
called from the main wrapper helper script /usr/bin/thunderbird.

Also adding some extra wrapper option to get the fixing for the MIME
part also available ouside the profile adption.

The logic for the adoption is hopefully simple.
If there is no folder or symlink ~/.thunderbird existing we can simply
create the link to ~/.icedove.
Otherwise we check if an existing symlink ~/.thunderbird isn't pointing
to ~/.icedove and error out then.

Closes: #855265, #855391, #855501